Revisions to REPL
Changes necessary to make basic REPL functionality work.

Major refactoing: Code of Interpreter is now in CompilingInterpreter.scala.
Interpreter.scala contains just the API.
