--- src/dotty/tools/dotc/repl/InterpreterResults.scala
@@ -1,19 +1,0 @@
-package dotty.tools.dotc.repl
-
-object InterpreterResults {
-
-  /** A result from interpreting one line of input. */
-  abstract sealed class Result
-
-  /** The line was interpreted successfully. */
-  case object Success extends Result
-
-  /** The line was erroneous in some way. */
-  case object Error extends Result
-
-  /** The input was incomplete.  The caller should request more
-   *  input.
-   */
-  case object Incomplete extends Result
-
-}
